==> CODE/00-Google_Earth_Engine.R <==
## First analysis step is to run spatial analyses on 
## Google Earth Engine (GEE). Link to code below. The files output from GEE have filenames
## following the pattern "heterogeneity_results_xxx_meters.csv" , where xxx is the 
## spatial grain of analysis in meters
## These output files are processed in the remainder of the analysis pipeline
## https://code.earthengine.google.com/7af120f2d4387b98435d90c423f74a9e
